[ALLUXIO-1910] Move Version to alluxio.cli package
