+++ core/server/src/main/java/alluxio/master/AlluxioMaster.java
@@ -15,7 +15,7 @@ import alluxio.AlluxioURI;
 import alluxio.Configuration;
 import alluxio.Constants;
 import alluxio.cli.ValidateConf;
-import alluxio.Version;
+import alluxio.cli.Version;
 import alluxio.master.block.BlockMaster;
 import alluxio.master.file.FileSystemMaster;
 import alluxio.master.journal.ReadWriteJournal;
