+++ core/server/src/main/java/alluxio/master/AlluxioMasterRestServiceHandler.java
@@ -15,7 +15,7 @@ import alluxio.Configuration;
 import alluxio.Constants;
 import alluxio.MasterStorageTierAssoc;
 import alluxio.RestUtils;
-import alluxio.Version;
+import alluxio.cli.Version;
 import alluxio.master.block.BlockMaster;
 import alluxio.underfs.UnderFileSystem;
 import alluxio.util.CommonUtils;
