+++ core/server/src/main/java/alluxio/web/WebInterfaceWorkerGeneralServlet.java
@@ -12,7 +12,7 @@
 package alluxio.web;
 
 import alluxio.Constants;
-import alluxio.Version;
+import alluxio.cli.Version;
 import alluxio.collections.Pair;
 import alluxio.util.FormatUtils;
 import alluxio.worker.WorkerContext;
