+++ examples/src/main/java/alluxio/examples/keyvalue/KeyValueStoreOperations.java
@@ -14,7 +14,7 @@ package alluxio.examples.keyvalue;
 import alluxio.AlluxioURI;
 import alluxio.Configuration;
 import alluxio.Constants;
-import alluxio.Version;
+import alluxio.cli.Version;
 import alluxio.client.ClientContext;
 import alluxio.client.keyvalue.KeyValueIterator;
 import alluxio.client.keyvalue.KeyValuePair;
