+++ examples/src/main/java/alluxio/examples/keyvalue/SameKeyValueStoresTest.java
@@ -13,7 +13,7 @@ package alluxio.examples.keyvalue;
 
 import alluxio.AlluxioURI;
 import alluxio.Constants;
-import alluxio.Version;
+import alluxio.cli.Version;
 import alluxio.client.keyvalue.KeyValueIterator;
 import alluxio.client.keyvalue.KeyValuePair;
 import alluxio.client.keyvalue.KeyValueStoreReader;
