+++ tests/src/test/java/alluxio/master/AlluxioMasterRestApiTest.java
@@ -15,7 +15,7 @@ import alluxio.Configuration;
 import alluxio.Constants;
 import alluxio.LocalAlluxioClusterResource;
 import alluxio.MasterStorageTierAssoc;
-import alluxio.Version;
+import alluxio.cli.Version;
 import alluxio.WorkerStorageTierAssoc;
 import alluxio.master.block.BlockMaster;
 import alluxio.metrics.MetricsSystem;
